refactor: 오타 수정으로 인한 패키지명 정정 (caculator → calculator)
