+++ src/main/java/calculator/PositiveOrZero.java
@@ -1,4 +1,4 @@
-package caculator;
+package calculator;
 
 import java.util.Objects;
