+++ src/main/java/calculator/PositiveOrZeros.java
@@ -1,4 +1,4 @@
-package caculator;
+package calculator;
 
 import java.util.Arrays;
 import java.util.List;
