+++ src/main/java/calculator/StringCalculator.java
@@ -1,4 +1,4 @@
-package caculator;
+package calculator;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
