+++ src/test/java/calculator/PositiveOrZeroTest.java
@@ -1,4 +1,4 @@
-package caculator;
+package calculator;
 
 import static org.assertj.core.api.Assertions.*;
